Split crate into features
