+++ Cargo.toml
@@ -8,7 +8,17 @@ license = "MIT OR Apache-2.0"
 authors = ["Owen Griffiths <root@ogriffiths.com>"]
 edition = "2021"
 
+[package.metadata.docs.rs]
+all-features = true
+rustdoc-args = ["--cfg", "docsrs"]
+
 [dependencies]
 axum = "0.5"
 http = "0.2"
 base64 = "0.13"
+
+[features]
+auth-basic = []
+auth-bearer = []
+
+default = ["auth-basic", "auth-bearer"]
